made Atom a child of Atomproperty
This was necessary to be able to call atom.GetAtomId() and so on, as with
C++ code and with old Boost::Python bindings. However there is an issue with
thisptr which is already defined as an Atomproperty* pointer. The workaround
is to first deallocate the Atomproperty memory and then allocate a new Atom.
This looks slightly inefficient and maybe a better solution will be found in
the future.